+++ src/main/java/com/bv/pet/jeduler/entities/Subtask.java
@@ -1,6 +1,5 @@
 package com.bv.pet.jeduler.entities;
 
-import com.bv.pet.jeduler.entities.user.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -16,7 +15,7 @@ import lombok.Setter;
 public class Subtask {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
+    private Integer id;
 
     @ManyToOne
     @JoinColumn(name = "task_id")
