+++ src/main/java/com/bv/pet/jeduler/repositories/NotificationRepository.java
@@ -3,5 +3,5 @@ package com.bv.pet.jeduler.repositories;
 import com.bv.pet.jeduler.entities.Notification;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NotificationRepository extends JpaRepository<Notification, Long> {
+public interface NotificationRepository extends JpaRepository<Notification, Integer> {
 }
