+++ src/main/java/com/bv/pet/jeduler/repositories/SubtaskRepository.java
@@ -3,5 +3,5 @@ package com.bv.pet.jeduler.repositories;
 import com.bv.pet.jeduler.entities.Subtask;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SubtaskRepository extends JpaRepository<Subtask, Long> {
+public interface SubtaskRepository extends JpaRepository<Subtask, Integer> {
 }
